Modified and tested the code to allow for proper creation/deletion of the patch data used with the pixie code.
Added a PatchContainer class that creates and destroys this data for a single patch (previously this was handled by the LevelContainer class).
Moved the routines to create the equilibrium, the grid structures, and the setup for the initialization of a patch object to the PatchContainer class.  This better reflects the operation of the code since these routines need to be called any time a new patch is created, not just when setting the initial conditions.  Any data in u is not overwritten.  The initial conditions now explicitly copies the data from u_0 to u to reflect this change.
The LevelContainer class was simplified to create and destroy the PatchContainer.  It will need to be updated/replaced with a container to create and destroy the PatchContainers when new levels are created or a regrid occurs.
The code was tested creating and destroying the PatchContainer object frequently during use without affecting the simulation.  This helped to remove many of the dependencies.

Additional code was added to the SAMRADriver to create multiple levels.  This has not been tested with more than one level yet.