Modified pixied3dApplication to cache the refine schedules.  This involved inheriting from mesh::StandardTagAndInitStrategy<NDIM>.  Currently I have tested the cached schedules using a static grid on the coarse level only.
Still need to add coarsen schedules and test multi-level.

Traced the code and performed some optimizations.  Currently, the time is dominated by recursiveFill in SAMRAI (for the small test runs performed).

Added support for a debug file which writes a single coarse grid to a binary file for direct comparisons.  When run, the resulting file will produce exactly the same results when run with 1 or more patches and can be checked using the diff command.  Added support for writing the debug files to the input file using 2 new arguments:
   save_debug   0: Do not write debug files (default), 1: write a debug file containing all ghosts cells and all patches, 2: write a debug file that contains only the coarse cells (suitable for direct comparisons using diff)
   debug_name   Name of the debug file to write (a default name of "debugFile" will be used if this is not specified)