Enable ADIOS2 in CMake
Add option to cmake for ADIOS v2. This allows the framework now to be
build with ADIOS2 support. Currently only available with parallel
compiliation. ADIOS v1 can still be used per CMake option.

Dependency building scripts are updated to compile ADIOS2 in addition to
all previous packages. The layout of the dependencies was changed to
compile them for serial or parallel, so the prefix path must be set
correctly when the build folder is initialized.